refactor commit info
